:cake: Fixed Limit Switch
So, make sure that you have the right port for you limit switch or it
won't work. Apparently the limit switches belong on the digital port
even though they are analog. Also, you don't need to set the pin mode
for a limit switch. If you don't think you need a line of code, try it
without and see what happens.